Added ability to set index for primary dapps.
If user is trying to set an index for some dapp1 and there is some other dapp2 which index is the same as the new one, app will ask him if he wants to overwrite dapp2 index.
Overwriting means that dapp's index will be set to -1, which is not a valid index, so the dapp will be sorted by the date it was created.