Updated methods for downloading Dapps.
Secondary and unapproved Dapps will be fully downloaded no matter how many of them there are. It used to be that app could only download first 1000 Dapps, since it's the maximum number of objects per request.
Adding Dapps to primary array is now restricted by the max number of 200 elements.